requirements_test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements_test.in
#
alabaster==0.7.12
    # via sphinx
alembic==1.4.2
    # via -r requirements_test.in
alembic-autogenerate-enums==0.0.2
    # via -r requirements_test.in
argh==0.26.2
    # via watchdog
atomicwrites==1.3.0
    # via pytest
attrs==20.3.0
    # via
    #   -r requirements_test.in
    #   pytest
babel==2.8.0
    # via sphinx
backcall==0.1.0
    # via ipython
beautifulsoup4==4.9.3
    # via -r requirements_test.in
bumpversion==0.5.3
    # via -r requirements_test.in
certifi==2019.11.28
    # via requests
chardet==3.0.4
    # via requests
colorama==0.4.3
    # via pytest-watch
coverage==5.0.3
    # via pytest-cov
decorator==4.4.1
    # via
    #   ipython
    #   traitlets
docopt==0.6.2
    # via pytest-watch
docutils==0.15.2
    # via
    #   -r requirements_test.in
    #   sphinx
factory-boy==3.2.0
    # via -r requirements_test.in
faker==8.0.0
    # via
    #   -r requirements_test.in
    #   factory-boy
flake8==3.9.0
    # via -r requirements_test.in
fpdf==1.7.2
    # via -r requirements_test.in
freezegun==0.1.11
    # via -r requirements_test.in
gunicorn==20.1.0
    # via -r requirements_test.in
httpretty==1.0.2
    # via -r requirements_test.in
idna==2.8
    # via requests
imagesize==1.2.0
    # via sphinx
ipdb==0.11
    # via -r requirements_test.in
ipython==7.12.0
    # via ipdb
ipython-genutils==0.2.0
    # via traitlets
jedi==0.16.0
    # via ipython
jinja2==2.11.1
    # via sphinx
lxml==4.6.3
    # via -r requirements_test.in
mako==1.1.2
    # via alembic
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
mccabe==0.6.1
    # via flake8
meld3==2.0.1
    # via supervisor
more-itertools==8.2.0
    # via pytest
packaging==20.1
    # via sphinx
parso==0.6.1
    # via jedi
pathtools==0.1.2
    # via watchdog
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pika==0.13
    # via pyamqp
pluggy==0.13.1
    # via pytest
pockets==0.9.1
    # via sphinxcontrib-napoleon
prompt-toolkit==3.0.3
    # via ipython
psycopg2-binary==2.8.3
    # via -r requirements_test.in
ptyprocess==0.6.0
    # via pexpect
py==1.8.1
    # via pytest
pyamqp==0.0.8.1
    # via -r requirements_test.in
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.5.2
    # via
    #   ipython
    #   sphinx
pymemcache==2.1.1
    # via -r requirements_test.in
pyparsing==2.4.6
    # via packaging
pytest==3.8.0
    # via
    #   -r requirements_test.in
    #   pyamqp
    #   pytest-cov
    #   pytest-notifier
    #   pytest-watch
pytest-cov==2.6.0
    # via -r requirements_test.in
pytest-notifier==1.0.1
    # via -r requirements_test.in
pytest-watch==4.2.0
    # via -r requirements_test.in
python-dateutil==2.8.1
    # via
    #   alembic
    #   faker
    #   freezegun
python-editor==1.0.4
    # via alembic
pytidylib==0.3.2
    # via -r requirements_test.in
pytz==2019.3
    # via babel
pyyaml==5.4.1
    # via watchdog
redis==3.2.1
    # via -r requirements_test.in
requests==2.22.0
    # via sphinx
six==1.14.0
    # via
    #   packaging
    #   pockets
    #   pymemcache
    #   pytest
    #   python-dateutil
    #   sphinxcontrib-httpdomain
    #   sphinxcontrib-napoleon
    #   traitlets
snowballstemmer==2.0.0
    # via sphinx
soupsieve==2.2.1
    # via beautifulsoup4
sphinx==2.1.2
    # via
    #   -r requirements_test.in
    #   sphinx-rtd-theme
    #   sphinxcontrib-httpdomain
    #   sphinxcontrib-plantuml
sphinx-rtd-theme==0.4.3
    # via -r requirements_test.in
sphinxcontrib-applehelp==1.0.1
    # via sphinx
sphinxcontrib-devhelp==1.0.1
    # via sphinx
sphinxcontrib-htmlhelp==1.0.2
    # via sphinx
sphinxcontrib-httpdomain==1.7.0
    # via -r requirements_test.in
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-napoleon==0.7
    # via -r requirements_test.in
sphinxcontrib-plantuml==0.20.1
    # via -r requirements_test.in
sphinxcontrib-qthelp==1.0.2
    # via sphinx
sphinxcontrib-serializinghtml==1.1.3
    # via sphinx
sqlalchemy==1.3.5
    # via
    #   -r requirements_test.in
    #   alembic
supervisor==4.0.4
    # via -r requirements_test.in
text-unidecode==1.3
    # via faker
traitlets==4.3.3
    # via ipython
urllib3==1.24.3
    # via
    #   -r requirements_test.in
    #   requests
watchdog==0.9.0
    # via
    #   -r requirements_test.in
    #   pytest-watch
wcwidth==0.1.8
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
